# fix(@angular-devkit/build-angular): add debugging and timing information in JavaScript and CSS optimization plugins